refactor: split skill rest service tests
